Move from TypedArrays to new BinArrays
TypedArrays will not allow offsets that are not multiples of
the array element size. This is a big problem when trying to
create structures like the following:

someStruct = new Struct([
  "uint8 a",
  "uint32 b[2]"
])

In order to create a TypeArray for the array of 32-bit ints the
offset has to be a multiple of 4 bytes. However, the first field
is only a single byte. This results in an error when creating the
TypeArray.

The solution was to create a new set of array like objects called
BinArrays. BinArrays do not have the same offset limitations as
TypeArrays and are instead implemented internally using DataViews.